+++ deploy/data/php/root/var/www/html/views/site/index.php
@@ -1,0 +1,30 @@
+<?php
+
+/** @var yii\web\View $this */
+
+$this->title = 'My Yii Application';
+?>
+<div class="site-index">
+
+    <div class="jumbotron text-center bg-transparent">
+        <h1 class="display-4">Congratulations!</h1>
+
+        <p class="lead">You have successfully created your Yii-powered application.</p>
+
+        <p><a class="btn btn-lg btn-success" href="http://www.yiiframework.com">Get started with Yii</a></p>
+    </div>
+
+    <div class="body-content">
+
+        <div class="row">
+            <div class="col-lg-4">
+                <h2>yii2-beter-logging</h2>
+
+                <p>A set of tests to debug or develop library.</p>
+
+                <p><a class="btn btn-outline-secondary" href="<?php echo \yii\helpers\Url::to(['beter-logging/index']) ?>">Test yii2-beter-logging</a></p>
+            </div>
+        </div>
+
+    </div>
+</div>